requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
awsebcli==3.15.3
beautifulsoup4==4.7.1
blessed==1.15.0
botocore==1.12.200
bs4==0.0.1
CacheControl==0.12.5
cached-property==1.5.1
cachetools==3.1.1
cement==2.8.2
certifi==2019.6.16
chardet==3.0.4
chromedriver-binary==76.0.3809.126.0
Click==7.0
colorama==0.3.9
discord.py==1.2.3
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
facebook-sdk==3.1.0
firebase-admin==2.17.0
Flask==1.1.0
future==0.16.0
google-api-core==1.13.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.2
google-cloud-firestore==1.2.0
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.22.0
httplib2==0.13.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==2.6.0
MarkupSafe==1.1.1
msgpack==0.6.1
multidict==4.5.2
oauthlib==3.1.0
pathspec==0.5.9
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.2.0
rsa==4.0
selenium==3.141.0
semantic-version==2.5.0
six==1.11.0
soupsieve==1.9.2
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
websockets==6.0
Werkzeug==0.15.4
yarl==1.3.0
